requirement.txt
 streamlit==1.44.1
 psycopg2-binary==2.0.3
 pandas==2.9.10
 pydeck==0.9.1